refactor: split profilePageController
